move normalise geobox to query
